+++ components/faphub/catalogtab/impl/src/main/java/com/flipperdevices/faphub/catalogtab/impl/composable/categories/ComposableCategories.kt
@@ -7,10 +7,11 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyListScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.flipperdevices.core.ui.errors.ComposableThrowableError
 import com.flipperdevices.core.ui.ktx.clickableRipple
 import com.flipperdevices.faphub.catalogtab.impl.model.CategoriesLoadState
 import com.flipperdevices.faphub.dao.api.model.FapCategory
+import com.flipperdevices.faphub.errors.api.FapErrorSize
+import com.flipperdevices.faphub.errors.api.FapHubComposableErrorsRenderer
 
 private const val DEFAULT_CATEGORIES_SIZE = 12
 private const val COLUMN_COUNT = 3
@@ -19,7 +20,8 @@ private const val COLUMN_COUNT = 3
 fun LazyListScope.ComposableCategories(
     loadState: CategoriesLoadState,
     onCategoryClick: (FapCategory) -> Unit,
-    onRetry: () -> Unit
+    onRetry: () -> Unit,
+    errorsRenderer: FapHubComposableErrorsRenderer
 ) {
     when (loadState) {
         is CategoriesLoadState.Loaded -> ComposableCategoriesGridItems(
@@ -34,14 +36,15 @@ fun LazyListScope.ComposableCategories(
             onClick = onCategoryClick
         )
 
-        is CategoriesLoadState.Error -> item {
-            ComposableThrowableError(
+        is CategoriesLoadState.Error -> with(errorsRenderer) {
+            ComposableThrowableErrorListItem(
                 modifier = Modifier
                     .fillMaxWidth()
                     .padding(horizontal = 14.dp)
                     .height(height = 250.dp),
                 throwable = loadState.throwable,
-                onRetry = onRetry
+                onRetry = onRetry,
+                fapErrorSize = FapErrorSize.IN_LIST
             )
         }
     }