+++ components/faphub/category/impl/src/main/java/com/flipperdevices/faphub/category/impl/composable/ComposableFapHubCategoryScreen.kt
@@ -6,6 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
+import androidx.paging.LoadState
 import androidx.paging.compose.collectAsLazyPagingItems
 import com.flipperdevices.core.ui.ktx.OrangeAppBarWithIcon
 import com.flipperdevices.core.ui.ktx.elements.SwipeRefresh
@@ -13,6 +14,8 @@ import com.flipperdevices.faphub.appcard.composable.paging.ComposableFapsList
 import com.flipperdevices.faphub.appcard.composable.paging.ComposableSortChoice
 import com.flipperdevices.faphub.category.impl.viewmodel.FapHubCategoryViewModel
 import com.flipperdevices.faphub.dao.api.model.FapItemShort
+import com.flipperdevices.faphub.errors.api.FapErrorSize
+import com.flipperdevices.faphub.errors.api.FapHubComposableErrorsRenderer
 import tangle.viewmodel.compose.tangleViewModel
 import com.flipperdevices.core.ui.res.R as DesignSystem
 
@@ -21,6 +24,7 @@ fun ComposableFapHubCategory(
     onBack: () -> Unit,
     onOpenSearch: () -> Unit,
     onOpenFapItem: (FapItemShort) -> Unit,
+    errorsRenderer: FapHubComposableErrorsRenderer,
     modifier: Modifier = Modifier,
     installationButton: @Composable (FapItemShort?, Modifier) -> Unit
 ) {
@@ -34,17 +38,25 @@ fun ComposableFapHubCategory(
             endIconId = DesignSystem.drawable.ic_search,
             onEndClick = onOpenSearch
         )
-
         SwipeRefresh(onRefresh = fapsList::refresh) {
             LazyColumn {
-                item {
-                    ComposableSortChoice(
-                        title = null,
-                        sortType = sortType,
-                        onSelectSortType = categoryViewModel::onSelectSortType
-                    )
+                if (fapsList.loadState.refresh !is LoadState.Error) {
+                    item {
+                        ComposableSortChoice(
+                            title = null,
+                            sortType = sortType,
+                            onSelectSortType = categoryViewModel::onSelectSortType
+                        )
+                    }
                 }
-                ComposableFapsList(fapsList, onOpenFapItem, installationButton)
+
+                ComposableFapsList(
+                    faps = fapsList,
+                    onOpenFapItem = onOpenFapItem,
+                    installationButton = installationButton,
+                    errorsRenderer = errorsRenderer,
+                    defaultFapErrorSize = FapErrorSize.FULLSCREEN
+                )
             }
         }
     }