+++ components/faphub/errors/api/src/main/java/com/flipperdevices/faphub/errors/api/FapErrorSize.kt
@@ -1,0 +1,9 @@
+package com.flipperdevices.faphub.errors.api
+
+import androidx.compose.ui.unit.TextUnit
+import androidx.compose.ui.unit.sp
+
+enum class FapErrorSize(val textSize: TextUnit) {
+    IN_LIST(12.sp),
+    FULLSCREEN(14.sp)
+}