+++ components/faphub/errors/api/src/main/java/com/flipperdevices/faphub/errors/api/throwable/FlipperNotConnected.kt
@@ -1,0 +1,3 @@
+package com.flipperdevices.faphub.errors.api.throwable
+
+class FlipperNotConnected : Throwable()