+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [FIX] Custom install and file manager uploader, deeplink refactor
 - [FIX] FapHub build status hide icon instead of info
 - [Feature] Add pull to refresh to fapscreen and category screen
+- [Feature] Add "Not Compatible with your Firmware" error
 
 # 1.6.0